Extract contracts into `Contracts` project
