+++ WineMate.Catalog/Features/WineMakers/CreateWineMaker.cs
@@ -8,11 +8,12 @@ using Mapster;
 
 using MediatR;
 
-using WineMate.Catalog.Contracts;
 using WineMate.Catalog.Database;
 using WineMate.Catalog.Database.Entities;
 using WineMate.Catalog.Extensions;
 using WineMate.Catalog.Validators;
+using WineMate.Contracts.Api;
+using WineMate.Contracts.Common;
 
 namespace WineMate.Catalog.Features.WineMakers;
