+++ WineMate.Catalog/Features/Wines/GetWine.cs
@@ -8,9 +8,9 @@ using MediatR;
 
 using Microsoft.EntityFrameworkCore;
 
-using WineMate.Catalog.Contracts;
 using WineMate.Catalog.Database;
 using WineMate.Catalog.Extensions;
+using WineMate.Contracts.Api;
 
 namespace WineMate.Catalog.Features.Wines;
