+++ WineMate.Catalog/WineMate.Catalog.csproj
@@ -41,4 +41,8 @@
         </Content>
     </ItemGroup>
 
+    <ItemGroup>
+        <ProjectReference Include="..\WineMate.Contracts\WineMate.Contracts.csproj"/>
+    </ItemGroup>
+
 </Project>
