+++ WineMate.Contracts/Common/WineType.cs
@@ -1,0 +1,12 @@
+namespace WineMate.Contracts.Common;
+
+public enum WineType
+{
+    Other,
+    Red,
+    White,
+    Rose,
+    Sparkling,
+    Dessert,
+    Fortified
+}
